refactor(storage): extract the links engine out of db.rs into storage/links.rs
Pure MOVE, zero behavior change. The link engine (semantic/wikilink
edges, the seal-cycle strip family, purge/rederive, kNN) is lifted
verbatim out of the 27k-line db.rs into a new storage/links.rs as an
inherent-impl split (impl Db in the sibling module) — db.rs shrinks
~2265 lines (27259 -> 25004).

Moved: migrate_links, resolve_wikilink, backlinks_* (indexed fast path +
meta readers), upsert_link_tx, index_wikilinks_for_source,
set_companion_link, rederive_{companion,inbound}_for_folder,
upsert/delete_manual_link, the sealed-neighbour strip family
(strip_sealed_neighbour_markers + the marker/title helpers),
purge_links_tx, clear/trim_*_suggestions_tx, item_centroid,
knn_items_visible, auto_link_semantic, link_by_id, accept/dismiss_link,
links_for_visible, collapse_manual_duplicate_edges,
link_endpoint_title_visible, the sealed-at-rest link probes, the
LINK_DECISION_KEEP const, and 3 #[cfg(test)] helpers.

The 28 links tests stay in db.rs's mod tests (they share mem_db/seed
helpers that construct Db from its private conn field — moving them would
force a fragile field promotion the refactor is meant to avoid). They
call the moved methods as inherent Db methods (cross-file visible) and
run unchanged; db.rs #[test] count is 250 before and after.

The ONLY db.rs logic-adjacent change is minimal pub(crate) promotions of
shared module-level helpers (map_err, visibility_clause,
doc_sealed_at_rest_tx, backlink_sort_key, lock, the two visibility
readers) so the sibling module can call them — no gate/seal/purge LOGIC
touched (purge_links_tx / strip_sealed_neighbour_markers_tx /
auto_link_semantic diff byte-identical vs the pre-move snapshot).

Verified: cargo test --lib links 43/0, seal 145/0, lock 175/0 (the
leak/seal-gate tests that would catch any gating change all pass), clippy
--lib -D warnings clean, MSRV grep clean. Full suite re-gated by CI.
Off trunk bbedc9f.
