+++ src-tauri/src/storage/mod.rs
@@ -1,4 +1,5 @@
 pub mod db;
+pub(crate) mod links;
 pub mod migration;
 pub mod models;
 pub mod usage;
